fix: Implement Routes and modified SessionRoutess
